+++ islands/EntryList/index.tsx
@@ -45,7 +45,6 @@ function entriesToContributions(entries: Document<LargeKvEntry>[]) {
 export function EntryList(props: iEntryList) {
   const [entries, setEntries] = useState<Document<LargeKvEntry>[]>([]);
   const [firstEntry, setFirstEntry] = useState<Document<LargeKvEntry>>();
-  const [lastEntry, setLastEntry] = useState<Document<LargeKvEntry>>();
 
   useEffect(() => {
     bring<findReq, Document<LargeKvEntry>[]>('/api/entries/find', 'POST', {
@@ -59,18 +58,12 @@ export function EntryList(props: iEntryList) {
             const currentDate = new Date(current.value.utc_created_at);
             return currentDate < earliestDate ? current : earliest;
           }, res[0]));
-          setLastEntry(res.reduce((latest, current) => {
-            const latestDate = new Date(latest.value.utc_created_at);
-            const currentDate = new Date(current.value.utc_created_at);
-            return currentDate > latestDate ? current : latest;
-          }, res[0]));
         }
       });
   }, [updateEntryList.value]);
 
   if (
     entries.length === 0 || firstEntry === undefined ||
-    lastEntry === undefined ||
     !firstEntry.value.utc_created_at === undefined
   ) {
     return <></>;
@@ -83,10 +76,7 @@ export function EntryList(props: iEntryList) {
           <ContributionCalendar
             contributionMap={entriesToContributions(entries)}
             startDateUtc={firstEntry.value.utc_created_at}
-            endDateUtc={lastEntry.value.utc_created_at}
-            // endDateUtc={new Date(
-            //   new Date().setDate(new Date().getDate() + 100),
-            // ).toUTCString()}
+            endDateUtc={new Date().toUTCString()}
           />
         )
         : <></>}