+++ routes/_app.tsx
@@ -1,3 +1,4 @@
+import { Stylesheet } from 'lunchbox';
 import { AppProps } from '$fresh/server.ts';
 import { Head } from '$fresh/runtime.ts';
 
@@ -6,10 +7,7 @@ export default function App({ Component }: AppProps) {
     <>
       <Head>
         <title>remember.land</title>
-        <link
-          rel='stylesheet'
-          href='https://cdn.jsdelivr.net/gh/CarcajadaArtificial/lunchbox@0.2.6/static/style.css'
-        />
+        <Stylesheet />
         <link rel='stylesheet' href='style.css' />
       </Head>
       <body class='clr-bg-panel clr-txt-base txt-paragraph'>